created function swithChar
created switchChar function in library file. Using RegExp to create a
dynamic pattern based on the function's argument, it will globally
replace all of a given character with another given character in a
given list.

added switchChar test function to main assignment file and edited the
test of the stringToNumber function.

edited content of sdi-project4.html